checking support for xml/uuid types. Renaming of test cases.
