+++ core/src/test/kotlin/org/evomaster/core/database/extract/postgres/DatetimeTypesTest.kt
@@ -5,22 +5,23 @@ import org.evomaster.client.java.controller.db.SqlScriptRunner
 import org.evomaster.client.java.controller.internal.db.SchemaExtractor
 import org.evomaster.core.database.DbActionTransformer
 import org.evomaster.core.database.SqlInsertBuilder
-import org.evomaster.core.search.gene.*
-import org.evomaster.core.search.gene.regex.RegexGene
-import org.evomaster.core.search.gene.sql.SqlAutoIncrementGene
+import org.evomaster.core.search.gene.datetime.DateGene
+import org.evomaster.core.search.gene.datetime.DateTimeGene
+import org.evomaster.core.search.gene.datetime.TimeIntervalGene
+import org.evomaster.core.search.gene.datetime.TimeGene
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 
 /**
  * Created by jgaleotti on 07-May-19.
  */
-class BinaryTypesExtractTest : ExtractTestBasePostgres() {
+class DatetimeTypesTest : ExtractTestBasePostgres() {
 
-    override fun getSchemaLocation() = "/sql_schema/postgres_binary_types.sql"
+    override fun getSchemaLocation() = "/sql_schema/postgres_datetime_types.sql"
 
 
     @Test
-    fun testBinaryTypes() {
+    fun testDatetimeTypes() {
 
         val schema = SchemaExtractor.extract(connection)
 
@@ -31,16 +32,26 @@ class BinaryTypesExtractTest : ExtractTestBasePostgres() {
 
         val builder = SqlInsertBuilder(schema)
         val actions = builder.createSqlInsertionAction(
-            "BinaryTypes", setOf(
-                "byteaColumn"
+            "DatetimeTypes", setOf(
+                "timestampColumn",
+                "timestampWithTimeZoneColumn",
+                "dateColumn",
+                "timeColumn",
+                "timeWithTimeZoneColumn",
+                "intervalColumn"
             )
         )
 
         val genes = actions[0].seeGenes()
 
-        assertEquals(1, genes.size)
-        assertTrue(genes[0] is StringGene) //character varying
-        assertEquals(0, (genes[0] as StringGene).minLength)
+        assertEquals(6, genes.size)
+
+        assertTrue(genes[0] is DateTimeGene)
+        assertTrue(genes[1] is DateTimeGene)
+        assertTrue(genes[2] is DateGene)
+        assertTrue(genes[3] is TimeGene)
+        assertTrue(genes[4] is TimeGene)
+        assertTrue(genes[5] is TimeIntervalGene)
 
         val dbCommandDto = DbActionTransformer.transform(actions)
         SqlScriptRunner.execInsert(connection, dbCommandDto.insertions)
