+++ core/src/test/kotlin/org/evomaster/core/database/extract/postgres/EnumeratedTypesTest.kt
@@ -1,0 +1,48 @@
+package org.evomaster.core.database.extract.postgres
+
+import org.evomaster.client.java.controller.api.dto.database.schema.DatabaseType
+import org.evomaster.client.java.controller.db.SqlScriptRunner
+import org.evomaster.client.java.controller.internal.db.SchemaExtractor
+import org.evomaster.core.database.DbActionTransformer
+import org.evomaster.core.database.SqlInsertBuilder
+import org.evomaster.core.search.gene.*
+import org.evomaster.core.search.gene.regex.RegexGene
+import org.evomaster.core.search.gene.sql.SqlAutoIncrementGene
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Test
+
+/**
+ * Created by jgaleotti on 07-May-19.
+ */
+class EnumeratedTypesTest : ExtractTestBasePostgres() {
+
+    override fun getSchemaLocation() = "/sql_schema/postgres_enumerated_types.sql"
+
+
+    @Test
+    fun testEnumeratedTypes() {
+
+        val schema = SchemaExtractor.extract(connection)
+
+        assertNotNull(schema)
+
+        assertEquals("public", schema.name.lowercase())
+        assertEquals(DatabaseType.POSTGRES, schema.databaseType)
+
+        val builder = SqlInsertBuilder(schema)
+        val actions = builder.createSqlInsertionAction(
+            "EnumeratedTypes", setOf(
+                "enumColumn"
+            )
+        )
+
+        val genes = actions[0].seeGenes()
+
+        assertEquals(1, genes.size)
+        assertTrue(genes[0] is EnumGene<*>)
+
+        val dbCommandDto = DbActionTransformer.transform(actions)
+        SqlScriptRunner.execInsert(connection, dbCommandDto.insertions)
+
+    }
+}
