+++ core/src/test/resources/sql_schema/postgres_uuid_type.sql
@@ -1,0 +1,4 @@
+CREATE TABLE UUIDType (
+           uuidColumn uuid NOT NULL
+);
+
