+++ core/src/test/resources/sql_schema/postgres_xml_type.sql
@@ -1,0 +1,4 @@
+CREATE TABLE XMLType (
+           xmlColumn xml NOT NULL
+);
+
